requirements.txt
# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5

# Search API
exa-py>=1.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web framework (for future API)
fastapi>=0.100.0
uvicorn>=0.23.0

# Memory / Storage
chromadb>=0.4.0
networkx>=3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# Optional: for enhanced features
# tiktoken>=0.5.0  # Token counting
# tenacity>=8.0.0  # Retry logic
